Base Window
I've created the base window for our game. What you should see when you run it are 5 red squares moving accross the screen from right to left at random heights. Currently, the enemy is part of the game class but i'm working on making it a class of its own.