Textures, Moved bullet x-origin by 10, Resizable window
I added textures to everything except the window background. These do not have to be in the final product in case anybody wants to draw original art instead of using images from google.

I made the window resizable, however the scaling is messed up if you drag the sides but fullscreen works perfectly. I will try to find a scaling fix.

I moved the bullet starting point to the right a bit since the old position was behind the character.

Issue: I don't know how to implement window drawing which is needed in order to create the "space.jpg" image for the background.